refactor: create Link directory
