+++ src/components/Link/ExternalLink/index.tsx
@@ -1,7 +1,7 @@
 import React, { FC, PropsWithChildren } from 'react';
 import styled from '@emotion/styled';
 import { RiExternalLinkFill } from '@remixicon/react';
-import Colors from '../../styles/Colors';
+import Colors from '../../../styles/Colors';
 
 interface ExternalLinkProps {
   readonly href: string;
